% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{DE_heatmap}
\alias{DE_heatmap}
\title{Draw DE Z-score heatmap for gene signatures}
\usage{
DE_heatmap(
  obj = NULL,
  sig_genes = NULL,
  type = c("standard", "hclust", "multiCCA"),
  direction = c("both", "down", "up"),
  top_n = 30,
  zscore_cap = 15,
  labRow = T,
  output_path = "./",
  prefix = "heatmap",
  height = 15,
  width = 12,
  ...
)
}
\arguments{
\item{obj}{the object produced by PCApermtest()}

\item{sig_genes}{the object produced by get_sig_genes()}

\item{type}{the type of obj and sig_genes that are being input. The "standard" (default) indicates 
the results from a standard within-perturbation analysis. The "hclust" indicates the results from 
a hierarchical clustering analysis. The "multiCCA" indicates the results from a multiCCA analysis.}

\item{direction}{to indicate whether gene signatures of both directions should be plotted, or just 
the up-regulated genes ("up") or down-regulated genes ("down") should be plotted.}

\item{top_n}{a positive integer to indicate how many gene signatures should be plotted. If provided,
the top top_n genes will be selected for plotting.}

\item{zscore_cap}{the cap for Z-scores in the Z-score matrix. Any Z-score that is larger than this 
value will be capped to this value.}

\item{labRow}{a boolen variable to indicate if the row names should be labelled in the heatmap or not.}

\item{output_path}{the directory where the heatmap will be saved.}

\item{prefix}{the prefix for how the file of the heatmap should be named}

\item{height}{the height (in inch) for the figure}

\item{width}{the width (in inch) for the figure}
}
\description{
This function will generate a standard heatmap based on the DE Z-score heatmap.
Only the selected significant gene signatures will be plotted in the rows.
}
