% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_test.R
\name{DEenrich}
\alias{DEenrich}
\title{Wrapper function for DE and enrichment test}
\usage{
DEenrich(
  object,
  plist = NULL,
  ident = NULL,
  ident.1 = NULL,
  ident.2 = NULL,
  ct.class = NULL,
  slct.ct = NULL,
  direction = c("up", "down", "both"),
  logfc.threshold = 0.25,
  p.val.cutoff = 0.05,
  min.pct = 0.1,
  assay = NULL,
  ...
)
}
\arguments{
\item{object}{a seurat object to perform the DE test and the enrichment test}

\item{plist}{the pathway gene lists to test the DE genes against}

\item{ident.1}{Identity class to define markers for; pass an object of class phylo or 'clustertree' to find markers for a node in a cluster tree; passing 'clustertree' requires BuildClusterTree to have been run}

\item{ident.2}{A second identity class for comparison; if NULL, use all other cells for comparison; if an object of class phylo or 'clustertree' is passed to ident.1, must pass a node to find markers for}

\item{ct.class}{Regroup cells into a different identity class prior to performing differential expression. 
Default is NULL (so all cells be used simultaneously).}

\item{slct.ct}{Subset a particular identity class prior to regrouping. Only relevant if group.by is set.}
}
\value{
a list of data frames containing the gene set enrichment results for each group in "group.by"
}
\description{
This function provides a wrapper of Seurat::FindMarkers() and Mixscale::fisher_enrich_test(). 
Users can input a Seurat object they want to investigate and a list of gene sets they want to 
test against, and the wrapper will perform DE tests + Fisher's enrichment test across all the 
available cell types. It will then return a list of data frames, containing gene set enrichment 
results for each cell type.
}
