% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{DEenrich_DotPlot}
\alias{DEenrich_DotPlot}
\title{Dot plot for enrichment results across multiple cell types}
\usage{
DEenrich_DotPlot(
  obj,
  adjust.methods = "BH",
  direction = c("up", "down", "both"),
  log10P_cutoff = 10,
  OR_cutoff = 20,
  slct_labels = NULL,
  plot_title = NULL
)
}
\arguments{
\item{obj}{the list of data frames generated by Mixscale_DEenrich()}

\item{adjust.methods}{the method for multiple testing correction method (see 'p.adjust.methods')}

\item{direction}{a character to specify what to plot: to plot the enrichment results for 
the up-regulated DE genes ("up") or the down-regulated DE genes ("down").}

\item{log10P_cutoff}{the maximum value of -log10(adjusted P-value) for plotting the size of the dot (any dot with a p-value 
larger than this will be set to the same size)}

\item{OR_cutoff}{the odds ratio (of the Fisher's exact test) cutoff for plotting the color gradient 
of the dot (any dot with a OR larger than this will be set to the same color)}

\item{slct_labels}{the selected labels (cell types) that need to be plotted}

\item{adjust.split}{TRUR/FALSE to specify if the multiple testing correction should be done for 
each group separately (TRUE) or all together (FALSE)}
}
\value{
a ggplot2 object
}
\description{
This function will generate a Dot plot for the enrichment results generated by Mixscale_DEenrich().
It will also perform multiple testing for the P-values for all the cell types taken together 
(or within each cell type).
}
