% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{DEmultiCCA}
\alias{DEmultiCCA}
\title{Run MultiCCA for a list of matrices}
\usage{
DEmultiCCA(
  mat_list = NULL,
  penalty = FALSE,
  standardize = FALSE,
  max_k = 5,
  cor_number = "all",
  mean_cor_thres = 0.2,
  flag_cor_num = T,
  flag_loose = F,
  pval_thres = 0.05,
  cor_coef_thres = 0.6,
  cor_coef_mean_thres = 0.3,
  ...
)
}
\arguments{
\item{mat_list}{the list of >= 2 DE Z-score matrices for the multiCCA analysis. Each matrix
should have the same named rows, but can have different number of columns (samples).}

\item{penalty}{to indicate if penalty should be applied during the multiCCA process. When set
to "FALSE", no penalty will be applied so that the results are not forced to be "sparse". However,
if a single value or a vector of k values (k should be the same as the number of matrices in the
input list), then L1 penalty will be applied to each matrix to force the output CVs to be sparse. 
See MultiCCA() from the PMA package for details.}

\item{standardize}{a boolen value to indicate whether to standardize each column before running 
the MultiCCA. Default is FALSE.}

\item{max_k}{the maximum number of MultiCCA runs. MultiCCA will be repeated until this number is 
reached or the CVs across the matrices have very low correlation coefficients.}

\item{mean_cor_thres}{During the multiCCA analysis, if any of the input matrix has significantly 
low(er) correlation with other matrices, it will impact the MultiCCA process. This is the threshold
to remove such matrices. If (the CV of) any matrix has a mean correlation coefficient <= 0.2, it will 
be removed and the MultiCCA will be repeated for this round (k-th). Note that this matrix will be appended 
back to the list for the next round (k+1-th) of MultiCCA, and the same filtering will be repeated. 
Default is 0.2. Set it to 0 to avoid such filtering. Vector is also accepted and will be sequentially used 
for each iteration of MultiCCA.}

\item{cor_num}{for each column of a matrix, the number of CVs that it needs to be significantly correlated 
with to be selected as a member of the program. Default is "all", meaning it needs to be significantly 
correlated with all other CVs. Alternatively, users may input a integer (e.g., 2, 3, ...).}
}
\value{
a list of MultiCCA results for each program identified.
}
\description{
A function to perform MultiCCA analysis (main function imported from package "PMA", 
see PMID 19377034 for details of the algorithm) that takes in a list of multiple 
Z-score matrices to find the canonical variates (CVs) that maximize the cross-matrices 
correlation. The MultiCCA process is modified so that it is not completed in one 
run when multiple rounds of CVs are desired. Instead, after each MultiCCA run, we will
identify the columns (samples) from each matrix that highly correlate with the CV of that 
matrix, and extract + remove them from the matrix. The next MultiCCA is performed to the 
list of such "filtered" matrices. This process is repeated until the desired number of 
runs is reached (set by users) or the CVs across the matrices have very low correlation 
coefficients.
}
\seealso{
[MultiCCA()]
}
