% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_test.R
\name{fisher_enrich_test}
\alias{fisher_enrich_test}
\title{Standard Fisher's exact test for enrichment analysis}
\usage{
fisher_enrich_test(
  input_list = NULL,
  background = NULL,
  go_term_db = NULL,
  list_gene = F,
  EASE = F
)
}
\arguments{
\item{input_list}{the input gene list}

\item{background}{the background gene list (usually the expressed genes where the 
input gene list is generate from, ).}

\item{go_term_db}{a list of gene-lists (GO term). It should be a list contain multiple named vector, 
and each vector should be a vector of multiple marker/signature genes for some biological pathway/process.}

\item{list_gene}{A Boolen value to indicate if the overlapping genes between the input gene list and 
the GO-term should be output as well.}

\item{EASE}{A Boolen value to indicate if the EASE correction should be applied (see 
https://david.ncifcrf.gov/helps/functional_annotation.html). This is useful to mitigate the 
small-sample inflation when the input gene list is short (e.g., < 10).}
}
\value{
a data frame contains the enrichment test results. Each row contains the P-value and enrichment odds
ratio calculated from a Fisher's exact test for one GO-term in the go_term_db.
}
\description{
This function will perform the strandard Fisher's exact test between the input gene 
list and a series of gene-ontology gene sets (adopted from DAVID GO analysis).
}
