% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fold_change.R
\name{FoldChange_new}
\alias{FoldChange_new}
\title{Calculate log-fold-change given a vector of gene expression and the indices of perturbed cells and non-target cells}
\usage{
FoldChange_new(
  object,
  cells.1,
  cells.2,
  mean.fxn,
  fc.name,
  features = NULL,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{cells.1}{Vector of cell names belonging to group 1}

\item{cells.2}{Vector of cell names belonging to group 2}

\item{mean.fxn}{Function to use for fold change or average difference calculation}

\item{fc.name}{Name of the fold change, average difference, or custom function column
in the output data.frame}

\item{features}{Features to calculate fold change for.
If NULL, use all features}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a single value of the log-fold-change of the input gene.
}
\description{
Function to calculate log-fold-change for pooled CRISPR screen datasets.
It is just a simple function to calculate the log-fold-change. Users can customise the min.cells, 
minimal expression threshold, pseudo-count (the small value added to the expression level to avoid log(0)), 
minimal percentage of cells expression the genes, and the base of the log.
}
\concept{perturbation_scoring}
