% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fold_change.R
\name{get_fc}
\alias{get_fc}
\title{Calculate log-fold-change given a vector of gene expression and the indices of perturbed cells and non-target cells}
\usage{
get_fc(
  gene_exp = NULL,
  idx_P = NULL,
  idx_NT = NULL,
  min.cells = 3,
  thresh.min = 0,
  pseudocount.use = 1,
  min.pct = 0.1,
  base = 2,
  norm.method = "raw"
)
}
\arguments{
\item{gene_exp}{a vector of the gene expression levels}

\item{idx_P}{a vector of index for the perturbed cells in the gene_exp}

\item{idx_NT}{a vector of index for the non-target cells (controls) in the gene_exp}

\item{min.cells}{the minimal number of cells that expresses the gene; if lower than this value, the 
fold-change will be returned as NA. Default is 3.}

\item{thresh.min}{the minimal value of expression; any expression value lower than this will be 
considered as 0. Default is 0.}

\item{pseudocount.use}{the small value that will be added to the log-transformation to avoid log(0).
For example, if a mean expression value is x, the final log-}

\item{min.pct}{the minimal proportion of cells in either groups that expresses the gene}

\item{base}{the base for log()}

\item{norm.method}{the normalization method for the input gene_exp. Default is 'raw', which means 
the original count value without normalization. The other supported values are 'log.norm', "scale.data". 
The mean.fxn() will change accordingly.}
}
\value{
Returns a single value of the log-fold-change of the input gene.
}
\description{
Function to calculate log-fold-change for pooled CRISPR screen datasets.
It is just a simple function to calculate the log-fold-change. Users can customise the min.cells, 
minimal expression threshold, pseudo-count (the small value added to the expression level to avoid log(0)), 
minimal percentage of cells expression the genes, and the base of the log.
}
\concept{perturbation_scoring}
