% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{get_sig_genes_DEmultiCCA}
\alias{get_sig_genes_DEmultiCCA}
\title{Run PCApermtest and get significant genes from DEmultiCCA}
\usage{
get_sig_genes_DEmultiCCA(
  obj = NULL,
  k = 1,
  var_prop = NULL,
  center = T,
  scale = T,
  num_iter = 200,
  row_filtering_pval = 0.05,
  var_prop_total = NULL,
  perm_pval_thres = 0.05,
  ori_pval_thres = 1.666667e-06,
  cor_threshold = 0.2,
  collapse = T,
  ...
)
}
\arguments{
\item{obj}{The results object produced by DEmultiCCA() function.}

\item{k}{the number of top PCs to extract. If set to NULL, the 
function will determine the number of PCs to extract based on 
the var_prop, which is the cut-off on the proportion of variance 
explained for each PC. Only PCs with prop_var greater than this 
value will be removed extracted.}

\item{var_prop}{if k is not set, then use this value as a cutoff for 
%var explained to select the PCs. Only the PCs with %var larger than this 
value will be selected (the top k PCs).}

\item{center}{a boolen value to indicate whether the column will be centered.}

\item{scale}{a boolen value to indicate whether the column will be scaled to 1.}

\item{num_iter}{the number of iteration for the permutation test.}

\item{var_prop_total}{similar to var_prop. The accumulated sum of the %var
of the top i = 1, 2, ... PCs will be calculated, and the top k PCs with the 
accumulated sum larger than this value to be selected.}

\item{perm_pval_thres}{the p-value threshold for the permutation test. Rows 
(genes) with permutation p-value lower then permtest_pval_thres or greater than 
(1 - permtest_pval_thres) will be selected. Default is 0.05.}

\item{ori_pval_thres}{the p-value threshold for the original DE test. In the original
Z-score matrix we store all the Z-score from the DE test. When selecting significant 
rows (genes), we might wish them to be significant in the original DE test as well. 
By setting this value, we force the rows (genes) to be both significant in the 
permutation test and the DE test. Default is 0.05/30000 = 1.666667e-06 (which is 
gene-wide significant after Bonferroni correction). Set it to 1 to avoid such filtering.}

\item{cor_threshold}{After the PCApermtest, the actual orientation of the PC
might not be the same as the orientation of its correlated columns in the original
matrix. We need to do correlation test between each PC and all the columns in the 
original matrix, and then use this cor_threshold to define and extract the 
correlated columns and then use them to determine the actual orientation of the PC.}

\item{collapse}{a boolen value to indicate when k >= 2, whether the significant 
genes from the top k PCs should be return together as one gene list (collapse = T) 
or separately for each k (collapse = F). Default is T.}
}
\value{
return a list of vectors, and each vector contains the signature genes identified for each MultiCCA program.
}
\description{
This function will use the output from the DEmultiCCA() and get the neccessary elements for PCApermtest and 
get the gene signatures for each perturbation program that DEmultiCCA() identifies. It works in a similar way 
as get_sig_genes_DEhclust() .
}
