% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_test.R
\name{Mixscale_DEenrich}
\alias{Mixscale_DEenrich}
\title{This script contains functions for enrichment test: 1. standard enrichment test method based on
Fisher's exact test, which looks at if the overlap between the input gene list and the go-term 
gene set is significantly larger than the overlap between the background control gene list and the go-term
gene set. 2. a new enrichment test method using the rank biased overlap (RBO, Webber et al. 2010), which not only quantifies the 
overlap between the input gene list and the go-term gene set, but also quantify if the rank of the 2 lists 
are concordant or not. It can be viewed as a weighted Jaccard index that puts decreasing weights to each 
rank. 
Wrapper function for DE and enrichment test}
\usage{
Mixscale_DEenrich(
  object,
  plist = NULL,
  labels = "cell_type",
  conditions = "treatment",
  ident.1 = NULL,
  ident.2 = NULL,
  direction = c("up", "down", "both"),
  logfc.threshold = 0.25,
  p.val.cutoff = 0.05,
  min.pct = 0.1,
  assay = NULL,
  ...
)
}
\arguments{
\item{object}{a seurat object to perform the DE test and the enrichment test}

\item{plist}{the pathway gene lists to test the DE genes against}

\item{labels}{the metadata column of cell type labels (or other annotations) to loop through}

\item{conditions}{the metadata column of the Identity for ident.1 and ident.2 to run DE test}

\item{ident.1}{Identity class to define markers for; pass an object of class phylo or 'clustertree' to find markers for a node in a cluster tree; passing 'clustertree' requires BuildClusterTree to have been run}

\item{ident.2}{A second identity class for comparison; if NULL, use all other cells for comparison; if an object of class phylo or 'clustertree' is passed to ident.1, must pass a node to find markers for}
}
\value{
a list of data frames containing the gene set enrichment results for each group in "labels"
}
\description{
This function provides a wrapper of Seurat::FindMarkers() and Mixscale::fisher_enrich_test(). 
Users can input a Seurat object they want to investigate and a list of gene sets they want to 
test against, and the wrapper will perform DE tests + Fisher's enrichment test across all the 
available cell types. It will then return a list of data frames, containing gene set enrichment 
results for each cell type.
}
