% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{Mixscale_DoHeatmap}
\alias{Mixscale_DoHeatmap}
\title{Single-cell heatmap for selected DE genes stratified by target expression}
\usage{
Mixscale_DoHeatmap(
  object = NULL,
  assay = "RNA",
  slot = "data",
  labels = "gene",
  nt.class.name = "NT",
  PRTB = NULL,
  slct_condition = "con1",
  slct_features = NULL,
  ...
)
}
\arguments{
\item{object}{a seurat object returned by RunMixscale()}

\item{assay}{the assay name to extract the expression level data from for plotting}

\item{slot}{the slot name to extract the expression level data from for plotting}

\item{labels}{the column name in the object's meta.data that contains the target
gene labels}

\item{nt.class.name}{the classification name of non-targeting gRNA cells}

\item{PRTB}{the gene name of the perturbation target to be plotted}

\item{slct_condition}{the name of the selected condition (e.g., cell type) to be 
plotted. If split.by was set to NULL during RunMixscale(), then user should leave it 
as the default value (i.e., "con1").}

\item{slct_features}{a vector of the names of the selected features (usually some DE 
genes) to be plotted in the heatmap.}
}
\value{
a ggplot2 object of the single-cell heatmap.
}
\description{
This function will generate single-cell expression heatmap for selected DE genes 
in cells of the same perturbation target (gRNA). This function is basically a 
wrapper function of the Seurat::DoHeatmap(), but with easier usage to select the 
cells based on given gRNA identity. Cells will be ordered in each stratification based 
on their perturbation scores.
}
