% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{Mixscale_RidgePlot}
\alias{Mixscale_RidgePlot}
\title{Plot to show the distribution of perturbation scores}
\usage{
Mixscale_RidgePlot(
  object = NULL,
  labels = "gene",
  nt.class.name = "NT",
  split.by = NULL,
  PRTB = NULL,
  slct_split.by = NULL,
  facet_wrap = c(NULL, "gene", "split.by"),
  facet_scale = c("fixed", "free_y"),
  facet_nrow = 1,
  ...
)
}
\arguments{
\item{object}{a seurat object returned by RunMixscale()}

\item{labels}{the column name in the object's meta.data that contains the target
gene labels}

\item{nt.class.name}{the classification name of non-targeting gRNA cells}

\item{split.by}{metadata column with experimental condition/cell type classification 
information. This is meant to be used to account for cases a perturbation is 
condition/cell type -specific.}

\item{PRTB}{the perturbation target genes to extract for plotting. Multiple values are 
allowed.}

\item{slct_split.by}{if only a subset of the conditions/cell-types in the split.by column need 
to be plotted, users can specify them as a character vector here. Default is NULL, meaning all the 
conditions/cell-types need to be plotted.}

\item{facet_wrap}{whether to divide the plot into multiple facets based on either the 
perturbation targets ("gene") or conditions/cell types ("split.by"). Default is NULL, meaning
no facet.}

\item{facet_scale}{whether to use a fixed scale for y-axis across all facets or allow 
y axis to vary.}

\item{facet_nrow}{the number of rows to plot the different panels when facet_wrap is set.}
}
\value{
a ggplot2 object that contains the ridge plot.
}
\description{
This function will generate a density (ridge) plot for the perturbation scores across
different cell types or different perturbation targets.
}
