% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{Mixscale_ScatterPlot}
\alias{Mixscale_ScatterPlot}
\title{Plot to compare the expression level of the perturbation target and the 
perturbation scores}
\usage{
Mixscale_ScatterPlot(
  object = NULL,
  assay = "RNA",
  slot = "data",
  labels = "gene",
  nt.class.name = "NT",
  split.by = NULL,
  PRTB = NULL,
  nbin = 10,
  facet_wrap = c(NULL, "gene", "split.by"),
  facet_scale = c("fixed", "free_y"),
  ...
)
}
\arguments{
\item{object}{a seurat object returned by RunMixscale()}

\item{assay}{the assay name to extract the expression level data from for plotting}

\item{slot}{the slot name to extract the expression level data from for plotting}

\item{labels}{the column name in the object's meta.data that contains the target
gene labels}

\item{nt.class.name}{the classification name of non-targeting gRNA cells}

\item{split.by}{metadata column with experimental condition/cell type classification 
information. This is meant to be used to account for cases a perturbation is 
condition/cell type -specific.}

\item{PRTB}{the perturbation target genes to extract for plotting. Multiple values are 
allowed.}

\item{nbin}{the number of bins to divide the perturbation scores into.}

\item{facet_wrap}{whether to divide the plot into multiple facets based on either the 
perturbation targets ("gene") or conditions/cell types ("split.by"). Default is NULL, meaning
no facet.}

\item{facet_scale}{whether to use a fixed scale for y-axis across all facets or allow 
y axis to vary.}
}
\value{
a ggplot2 object that contains the connected scatterplot.
}
\description{
This function will generate a connected scatterplot to compare the mean
expression level of the perturbation target gene within different perturbation 
percentile bins. After running the RunMixscale() function, user can specify the 
gene name of the perturbation target and the number of bins to divide the scores 
into, and this function will sutomatically generate a connected scatterplot. 
Multiple perturbation targets and cell types are allowed.
}
