% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{PCApermtest}
\alias{PCApermtest}
\title{Run PCA-based permutation test for a matrix}
\usage{
PCApermtest(
  mat = NULL,
  k = 1,
  var_prop = NULL,
  var_prop_total = NULL,
  center = T,
  scale = T,
  row_filtering_pval = 0.05,
  num_iter = 200,
  seed = 123124125,
  ...
)
}
\arguments{
\item{mat}{the Z-score matrix to perform the permutation test. 
Rows are the gene and columns are the conditions/samples.}

\item{k}{the number of top PCs to extract. If set to NULL, the 
function will determine the number of PCs to extract based on 
the var_prop, which is the cut-off on the proportion of variance 
explained for each PC. Only PCs with prop_var greater than this 
value will be removed extracted.}

\item{var_prop}{if k is not set, then use this value as a cutoff for 
%var explained to select the PCs. Only the PCs with %var larger than this 
value will be selected (the top k PCs).}

\item{var_prop_total}{similar to var_prop. The accumulated sum of the %var
of the top i = 1, 2, ... PCs will be calculated, and the top k PCs with the 
accumulated sum larger than this value to be selected.}

\item{center}{a boolen value to indicate whether the column will be centered.}

\item{scale}{a boolen value to indicate whether the column will be scaled to 1.}

\item{num_iter}{the number of iteration for the permutation test.}

\item{seed}{seed for random number generator.}
}
\value{
a list object consists of 3 elements: the original input Z-score matrix,
the p-value matrix produced by the permutation test for each genes (same dimension 
as the original input matrix), and the prcomp() object done for the input Z-score 
matrix.
}
\description{
This function will load in a DE test Z-score matrix, and perform PCA to it to get the 1st to k-th PCs. 
Then it will permuate the matrix and then the same PCA analysis will be performed. By default, this process
will be repeated for 200 times (default), so that we will have a decent number of null values 
in the top k PCs. The proportion of extreme values greater or smaller than the actual
value will be used as the P-values.
}
