% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PRTBscore_RidgePlot}
\alias{PRTBscore_RidgePlot}
\title{Plot to show the distribution of perturbation scores}
\usage{
PRTBscore_RidgePlot(
  object = NULL,
  labels = "gene",
  nt.class.name = "NT",
  split.by = NULL,
  PRTB = NULL,
  facet_wrap = c(NULL, "gene", "split.by"),
  facet_scale = c("fixed", "free_y"),
  ...
)
}
\arguments{
\item{object}{a seurat object returned by PRTBScoring()}

\item{labels}{the column name in the object's meta.data that contains the target
gene labels}

\item{nt.class.name}{the classification name of non-targeting gRNA cells}

\item{split.by}{metadata column with experimental condition/cell type classification 
information. This is meant to be used to account for cases a perturbation is 
condition/cell type -specific.}

\item{PRTB}{the perturbation target genes to extract for plotting. Multiple values are 
allowed.}

\item{facet_wrap}{whether to divide the plot into multiple facets based on either the 
perturbation targets ("gene") or conditions/cell types ("split.by"). Default is NULL, meaning
no facet.}

\item{facet_scale}{whether to use a fixed scale for y-axis across all facets or allow 
y axis to vary.}
}
\value{
a ggplot2 object that contains the ridge plot.
}
\description{
This function will generate a density (ridge) plot for the perturbation scores across
different cell types or different perturbation targets.
}
