% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation_scoring.R
\name{PRTBScoring}
\alias{PRTBScoring}
\title{Perturbation Scoring}
\usage{
PRTBScoring(
  object,
  assay = "PRTB",
  slot = "scale.data",
  labels = "gene",
  nt.class.name = "NT",
  new.class.name = "mixscape_class",
  min.de.genes = 5,
  min.cells = 5,
  de.assay = "RNA",
  logfc.threshold = 0.25,
  verbose = FALSE,
  split.by = NULL,
  fine.mode = FALSE,
  fine.mode.labels = "guide_ID",
  prtb.type = "KO",
  DE.gene = NULL,
  max.de.genes = 100,
  harmonize = F,
  min_prop_ntgd = 0.1,
  pval.cutoff = 0.05,
  seed = 10282021
)
}
\arguments{
\item{object}{An object of class Seurat.}

\item{assay}{Assay to use for mixscape classification.}

\item{slot}{Assay data slot to use.}

\item{labels}{metadata column with target gene labels.}

\item{nt.class.name}{Classification name of non-targeting gRNA cells.}

\item{new.class.name}{Name of mixscape classification to be stored in
metadata.}

\item{min.de.genes}{Required number of genes that are differentially
expressed for method to separate perturbed and non-perturbed cells.}

\item{min.cells}{Minimum number of cells in target gene class. If fewer than
this many cells are assigned to a target gene class during classification,
all are assigned NP.}

\item{de.assay}{Assay to use when performing differential expression analysis.
Usually RNA.}

\item{logfc.threshold}{the log-fold-change threshold to select the large-effect
DE genes. Only DE genes with log-fold-change larger than this value will be 
selected. Default is 0.25.}

\item{verbose}{Display messages}

\item{split.by}{metadata column with experimental condition/cell type
classification information. This is meant to be used to account for cases a
perturbation is condition/cell type -specific.}

\item{fine.mode}{When this is equal to TRUE, DE genes for each target gene
class will be calculated for each gRNA separately and pooled into one DE list
for calculating the perturbation score of every cell and their subsequent
classification.}

\item{fine.mode.labels}{metadata column with gRNA ID labels.}

\item{prtb.type}{specify type of CRISPR perturbation expected for labeling mixscape classifications. Default is KO.}

\item{DE.gene}{specify a list of user-defined large-effect DE genes to calculate the perturbation score.}

\item{max.de.genes}{the maximum number of top large-effect DE genes to calculate the perturbation score. Default is 100.}

\item{harmonize}{a boolen value to specify whether a harmonization of the cell-type proportion between the NT cells and 
the perturbed cells should be performed prior to the DE test. If fine.mode is TRUE, this harmonization step will be 
performed for each fine.mode gRNA. Default is FALSE.}

\item{min_prop_ntgd}{a minimal threshold to remove cells if any cell type has a proportion less than this value. It will
only be used when harmonize is TRUE. Default is 0.1.}

\item{pval.cutoff}{specify the DE test p-value cutoff (after Bonferroni correction) to select top large-effect DE genes.
Default is 0.05.}
}
\value{
Returns a Seurat object containing the perturbation scores. It is stored in the Tool Data of the object, not in the 
meta.data.
}
\description{
Function to calculate perturbation scores for perturbed and non-perturbed gRNA expressing cells.
The perturbation score reflects the perturbation strength of each cells (inherited from the RunMixscape()
function). It is calculated by using the large-effect DE genes from raw DE tests between the 
perturbed and non-perturbed gRNA expressing cells.
}
\concept{perturbation_scoring}
