% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_de.R
\name{prune_DE_mat}
\alias{prune_DE_mat}
\title{QC for the list of DE Z-score matrices}
\usage{
prune_DE_mat(
  DEG_mat = NULL,
  zscore_cap = 37,
  mask_target = FALSE,
  p_threshold = 0.05/30000,
  min_sig_DEG = 0,
  center = FALSE,
  scale = FALSE,
  ...
)
}
\arguments{
\item{DEG_mat}{the list of DE Z-score matrices that produced by get_DE_mat()}

\item{zscore_cap}{the cap value for the Z-scores. Any absolute(Z-score) larger than this value will be 
set to this value to avoid extreme values affecting the downstream analyses. Default is 37 
which is the machine precision limit for a Z-score to produce a non-zero P-value (= ~1e-300).}

\item{mask_target}{a boolen value to indicate if the Z-score of the perturbation target (labelled 
by the column names) should be masked as 0. Default is FALSE so no masking will happen.}

\item{p_threshold}{the P-value threshold to define the 'significant' DE genes. Default is 0.05/30000, which
is approximately the Bonferroni correction threshold for genome-wide DE tests (assuming 30,000 genes in total).}

\item{min_sig_DEG}{the minimal number of significant DE genes that each column must contain. Any 
column with sig DE genes less than this value will be removed from the matrix. Default is 0 so no 
column will be removed.}

\item{center}{a boolen value to indicate whether we should center each column to 0. Default is FALSE.}

\item{scale}{a boolen value to indicate whether we should scale each column to have variance = 1. Default is FALSE.}
}
\value{
The function will return a list of QCed DE Z-score matrices. This list can directly be the input for 
DEmultiCCA().
}
\description{
A function to perform QC/filtering for the DE Z-score matrices that produced by get_DE_mat().
}
