% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_test.R
\name{rbo}
\alias{rbo}
\title{This script contains functions for enrichment test: 1. standard enrichment test method based on
Fisher's exact test, which looks at if the overlap between the input gene list and the go-term 
gene set is significantly larger than the overlap between the background control gene list and the go-term
gene set. 2. a new enrichment test method using the rank biased overlap (RBO, Webber et al. 2010), which not only quantifies the 
overlap between the input gene list and the go-term gene set, but also quantify if the rank of the 2 lists 
are concordant or not. It can be viewed as a weighted Jaccard index that puts decreasing weights to each 
rank. 
Rank biased overlap}
\usage{
rbo(
  list1,
  list2,
  p,
  k = floor(max(length(list1), length(list2))/2),
  side = c("top", "bottom"),
  mid = NULL,
  uneven.lengths = TRUE
)
}
\arguments{
\item{list1}{List 1}

\item{list2}{List 2}

\item{p}{Weighting parameter in [0, 1]. High p implies strong emphasis on top ranked elements}

\item{k}{Evaluation depth for extrapolation}

\item{side}{Evaluate similarity between the top or the bottom of the ranked lists}

\item{mid}{Set the mid point to for example only consider positive or negative scores}

\item{uneven.lengths}{Indicator if lists have uneven lengths}
}
\value{
a scaler value measuring the rank biased overlap (rbo)
}
\description{
Functions for a new gene-set enrichment test based on the 
RBO (rank biased overlap) calculation with extropolation (Webber et al., 2010).
The core functions of rbo() calculation was modified from the "gespeR" package (original author: Fabian Schmich). 
We modified it to accomodate our package and data type. We also developed a permutation scheme for 
RBO to allow for p-value calculations.
}
\author{
Fabian Schmich ("gespeR" package)
}
