% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_de.R
\name{Run_wmvRegDE}
\alias{Run_wmvRegDE}
\title{Scoring-based weighted DE test}
\usage{
Run_wmvRegDE(
  object,
  assay = "RNA",
  slot = "counts",
  labels = "gene",
  nt.class.name = "NT",
  verbose = FALSE,
  PRTB_list = NULL,
  split.by = NULL,
  logfc.threshold = 0,
  min.pct = 0.1,
  min.cells.group = 10,
  total_ct_labels = "nCount_RNA",
  pseudocount.use = 1,
  base = 2,
  full.results = FALSE
)
}
\arguments{
\item{object}{An object of class Seurat.}

\item{assay}{Assay to use for mixscape classification.}

\item{slot}{Assay data slot to use.}

\item{labels}{metadata column with target gene labels.}

\item{nt.class.name}{Classification name of non-targeting gRNA cells.}

\item{verbose}{Print a progress bar once expression testing begins}

\item{PRTB_list}{provide a vector of perturbations that the DE tests are restricted 
to. Default is NULL (DE tests will be performed for all available perturbations).}

\item{split.by}{metadata column with experimental condition/cell type
classification information. This is used to account for cases where 
perturbations are done for multiple condition/cell type. Default is NULL (only one 
cell type).}

\item{logfc.threshold}{the log-fold-change threshold to select genes for DE
test. Genes with log2-fold-change larger than this value will be selected for DE test. 
Note that if split.by is set and more than 1 split.by group exists, this 
logfc.threashold will be applied to each group and if any of them satisfies this criteria, the 
gene will be selected. Default is 0 (no filtering based on log2-fold-change).}

\item{min.pct}{only test genes that are detected in a minimum fraction of min.pct cells in either 
of the two populations. Meant to speed up the function by not testing genes that are very 
infrequently expressed. Default is 0.1. Same as logfc.threshold, if split.by is set and more than 1 split.by 
group exists, thiswill be applied to each group and if any of them satisfies this criteria, the 
gene will be selected.}

\item{min.cells.group}{Minimum number of cells in one of the groups}

\item{total_ct_labels}{metadata column for the total RNA counts of each cell. The default is 
nCount_RNA, which is the default names used by Seurat package.}

\item{pseudocount.use}{Pseudocount to add to averaged expression values when calculating logFC. 1 by default.}

\item{base}{The base with respect to which logarithms are computed.}

\item{full.results}{A boolen value to indicate if the full DE results should be output. Default is 
FALSE (only the regression coefficients/P-values of the mixscale scores will be output).}
}
\value{
a list of DE results, one for each perturbation
}
\description{
A function to perform differential expression (DE) tests based on the perturbation scores from the 
RunMixscale() function. It is a multivariate negative binomial based model that incorporates both the heterogeneity 
of perturbation strength in each cell, as well as their cell type background.
}
\concept{perturbation_scoring}
