% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_de.R
\name{scoringDE}
\alias{scoringDE}
\title{Scoring-based DE test}
\usage{
scoringDE(
  object,
  assay = "PRTB",
  slot = "data",
  labels = "gene",
  nt.class.name = "NT",
  verbose = FALSE,
  split.by = "cell_type",
  total_ct_labels = "nCount_RNA",
  logfc.threshold = 0.2,
  pseudocount.use = 0.1,
  base = 2,
  min.pct = 0.1,
  min.cells = 10
)
}
\arguments{
\item{object}{An object of class Seurat.}

\item{assay}{Assay to use for mixscape classification.}

\item{slot}{Assay data slot to use.}

\item{labels}{metadata column with target gene labels.}

\item{nt.class.name}{Classification name of non-targeting gRNA cells.}

\item{verbose}{Print a progress bar once expression testing begins}

\item{split.by}{metadata column with experimental condition/cell type
classification information. This is meant to be used to account for cases a
perturbation is condition/cell type -specific.}

\item{total_ct_labels}{metadata column for the total RNA counts of each cell}

\item{logfc.threshold}{the log-fold-change threshold to select genes for DE
test. Genes with log-fold-change larger than this value will be selected for DE test. 
Note that if split.by is set and more than 1 split.by group exists, this 
logfc.threashold will be applied to each group. Default is 0.25.}

\item{pseudocount.use}{Pseudocount to add to averaged expression values when
calculating logFC. 0.1 by default.}

\item{base}{The base with respect to which logarithms are computed.}

\item{min.pct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations. Meant to speed up the function
by not testing genes that are very infrequently expressed. Default is 0.1}

\item{min.cells}{Minimum number of cells in target gene class. If fewer than
this many cells are assigned to a target gene class during classification,
all are assigned NP.}

\item{new.class.name}{Name of mixscape classification to be stored in
metadata.}

\item{min.de.genes}{Required number of genes that are differentially
expressed for method to separate perturbed and non-perturbed cells.}
}
\value{
list of DE results
}
\description{
Function to perform differential expression (DE) tests based on the perturbation scores from the 
PRTBScoring() function. It is a multivariate negative binomial based model that incorporates both the heterogeneity 
of perturbation strength in each cell, as well as their cell type background.
}
\concept{perturbation_scoring}
