% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{Mixscale_DoHeatmap}
\alias{Mixscale_DoHeatmap}
\title{Single-cell heatmap for selected DE genes stratified by target expression}
\usage{
Mixscale_DoHeatmap(
  object = NULL,
  assay = "RNA",
  slot = "data",
  labels = "gene",
  nt.class.name = "NT",
  slct.ident = NULL,
  mixscale.score.name = "mixscale_score",
  features = NULL,
  group.by = NULL,
  ct.class = NULL,
  slct.ct = NULL,
  ...
)
}
\arguments{
\item{object}{a seurat object returned by RunMixscale()}

\item{assay}{the assay name to extract the expression level data from for plotting}

\item{slot}{the slot name to extract the expression level data from for plotting}

\item{labels}{the column name in the object's meta.data that contains the target
gene labels}

\item{nt.class.name}{the classification name of non-targeting gRNA cells}

\item{slct.ident}{the name of the perturbation target in 'labels' to be plotted.}

\item{mixscale.score.name}{Name of mixscale scores to be stored in metadata. Default is "mixscale_score".}

\item{features}{A vector of features to plot.}

\item{group.by}{A vector of variables to group cells by; pass 'ident' to group by cell identity classes.
Default is the same as 'labels'.}

\item{ct.class}{the metadata colname that stores the cell type (or other conditions) information. 
If this is set and the correct slct.ct is provided, only cells with the corresponding slct.ct will 
be plotted. Default is NULL (so cells from all the ct.class will be plotted).}

\item{slct.ct}{a character of the group of cells in ct.class to be plotted. Default is NULL (so cells from all the ct.class will be plotted).}
}
\value{
a ggplot2 object of the single-cell heatmap.
}
\description{
This function will generate single-cell expression heatmap for selected DE genes 
in cells of the same perturbation target (gRNA). This function is basically a 
wrapper function of the Seurat::DoHeatmap(), but with easier usage to select the 
cells based on given gRNA identity. Cells will be ordered in each stratification based 
on their perturbation scores.
}
